chore: add test inputs.

+++ src/decremental_minimum_spanning_forest/helpful_classes/forest.hpp
@@ -27,7 +27,7 @@ class Forest{
 		*/
 		std::unordered_map<Key, Tree<Key> *> mapTrees; 
 
-		int totalWeight = 0;
+		long long totalWeight = 0;
 		std::unordered_set<Edge<int>, EdgeHash<int>> edges;
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
